feat: Implement 'Edit Pet' functionality
- Renamed AddPetScreen to SavePetScreen and modified it to support both adding and editing pets.
- In edit mode, SavePetScreen pre-fills with existing pet data (fetched if not passed directly) and calls updatePetProvider on save.
- Added an 'Edit' button to PetDetailScreen's AppBar.
- Updated AppRouter with a '/pet/:petId/edit' route that navigates to SavePetScreen, passing petId and the Pet object (as 'extra') for optimized editing.
- Ensured dynamic UI text (titles, button labels) and appropriate user feedback for both add and edit operations on SavePetScreen.

Users can now edit the basic details of their existing pets.
